requirement.txt
django-icons
django-mptt
django-easy_thumbnails